Reintroduce OpenSSL abstract class
In v2.x we had this class but to handle both RSA and ECDSA signers.
However the logic we had for ECDSA was incorrect and not following the
JWA specs.

Now it's back with kind of the same structure and we'll use something to
manipulate the ECDSA points to generate the expected value.

https://github.com/lcobucci/jwt/issues/30